feat(friend): 新增查询聊天记录接口，优化pom，提高mp与PageHelp的版本

+++ src/main/java/com/hmetao/code_dictionary/service/impl/UserServiceImpl.java
@@ -79,7 +79,7 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements Us
     @Override
     @Transactional
     public void registry(UserRegistryForm userRegistryForm) {
-        Integer count = baseMapper.selectCount(new LambdaQueryWrapper<User>()
+        Long count = baseMapper.selectCount(new LambdaQueryWrapper<User>()
                 .eq(User::getUsername, userRegistryForm.getUsername())
                 .or().eq(User::getMobile, userRegistryForm.getMobile()));
         if (count > 0) {
